Days 8 & 11: get rid of ugly pre-return peeking of answer
Sometimes, the answer doesn't fit the Int boundaries. String as return
type may be the most flexible and future-proof, but answer to each task
so far has been an integer number. So, migrate from Int to Long.

+++ src/main/kotlin/icu/trub/aoc/AbstractDay.kt
@@ -4,7 +4,7 @@ abstract class AbstractDay(val inputFileName: String) {
     /**
      * @return list of solutions, one per day part.
      */
-    fun solve(): List<Int?> = listOf(
+    fun solve(): List<Long?> = listOf(
         solvePartOne(),
         solvePartTwo()
     )
@@ -12,10 +12,10 @@ abstract class AbstractDay(val inputFileName: String) {
     /**
      * @return the solution for Part One of this day, if available, and `null` otherwise
      */
-    protected abstract fun solvePartOne(): Int?
+    protected abstract fun solvePartOne(): Long?
 
     /**
      * @return the solution for Part Two of this day, if available, and `null` otherwise
      */
-    protected abstract fun solvePartTwo(): Int?
+    protected abstract fun solvePartTwo(): Long?
 }
